refactor(whatsapp): extract attached socket session
