+++ extensions/whatsapp/src/monitor-inbox.streams-inbound-messages.test-support.ts
@@ -493,7 +493,7 @@ describe("web monitor inbox", () => {
     await listener.close();
   });
 
-  it("stays unavailable on connect in self-chat mode", async () => {
+  it("socket session stays unavailable on connect in self-chat mode", async () => {
     const { listener, sock } = await startInboxMonitor(vi.fn(async () => {}) as InboxOnMessage, {
       selfChatMode: true,
     });
@@ -854,7 +854,7 @@ describe("web monitor inbox", () => {
     await listener.close();
   });
 
-  it("uses a replacement socket for replies created before reconnect", async () => {
+  it("socket session uses a replacement socket for replies created before reconnect", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef: NonNullable<InboxMonitorOptions["socketRef"]> = { current: null };
 
@@ -941,7 +941,7 @@ describe("web monitor inbox", () => {
     await listener.close();
   });
 
-  it("waits for a replacement socket before sending replies", async () => {
+  it("socket session waits for a replacement socket before sending replies", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const { listener, sock, inbound } = await primeInboundReplyHandle({
@@ -1300,7 +1300,7 @@ describe("web monitor inbox", () => {
     );
   });
 
-  it("retries timed-out sends on the same socket without clearing the socket ref", async () => {
+  it("socket session retries timed-out sends without clearing the socket ref", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const { listener, sock, inbound } = await primeInboundReplyHandle({
@@ -1328,7 +1328,7 @@ describe("web monitor inbox", () => {
     await listener.close();
   });
 
-  it("rejects direct sends before Baileys sendMessage when reachout timelock is active", async () => {
+  it("socket session rejects direct sends while reachout timelock is active", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.fetchAccountReachoutTimelock.mockResolvedValueOnce({
@@ -1349,7 +1349,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("uses connection.update reachout timelock state before direct sends", async () => {
+  it("socket session uses connection updates before direct sends", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.ev.emit("connection.update", {
@@ -1371,7 +1371,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("allows direct sends after reachout timelock clears", async () => {
+  it("socket session allows direct sends after reachout timelock clears", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.ev.emit("connection.update", {
@@ -1396,7 +1396,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("refreshes inactive reachout timelock state before later direct sends", async () => {
+  it("socket session refreshes inactive reachout state before later direct sends", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.fetchAccountReachoutTimelock
@@ -1420,7 +1420,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("reuses a successful readiness preflight for the immediate direct send", async () => {
+  it("socket session reuses readiness preflight for the immediate direct send", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.fetchAccountReachoutTimelock.mockResolvedValueOnce({ isActive: false });
@@ -1436,7 +1436,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("invalidates readiness preflight when a later active timelock update arrives", async () => {
+  it("socket session invalidates readiness after an active timelock update", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.fetchAccountReachoutTimelock.mockResolvedValueOnce({ isActive: false });
@@ -1460,7 +1460,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("does not apply account reachout timelock to group sends", async () => {
+  it("socket session does not apply account reachout timelock to group sends", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     sock.ev.emit("connection.update", {
@@ -1480,7 +1480,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("blocks direct inbound composing presence when reachout timelock is active", async () => {
+  it("socket session blocks direct composing presence during reachout timelock", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const { listener, sock, inbound } = await primeInboundReplyHandle({
@@ -1507,7 +1507,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("times out stalled socket sends at the default Baileys query timeout", async () => {
+  it("socket session times out stalled sends at the Baileys query timeout", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     vi.useFakeTimers();
@@ -1524,7 +1524,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("does not retry or clear the socket after the local socket send timeout", async () => {
+  it("socket session preserves the socket after a local send timeout", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const { listener, sock, inbound } = await primeInboundReplyHandle({
@@ -1549,7 +1549,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("records outbound replies for Baileys retry lookup", async () => {
+  it("socket session records outbound replies for Baileys retry lookup", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const baileysCache = createBaileysCacheSupport();
@@ -1581,7 +1581,7 @@ describe("web monitor inbox", () => {
     await listener.close();
   });
 
-  it("suppresses self-echo when a timed-out socket send is later accepted", async () => {
+  it("socket session suppresses self-echo after a late accepted send", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const baileysCache = createBaileysCacheSupport();
@@ -1643,7 +1643,7 @@ describe("web monitor inbox", () => {
     await listener.close();
   });
 
-  it("times out stalled send-api presence updates", async () => {
+  it("socket session times out stalled send-api presence updates", async () => {
     const onMessage = vi.fn(async () => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage);
     vi.useFakeTimers();
@@ -1661,7 +1661,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("bounds stalled read-receipt socket operations instead of hanging", async () => {
+  it("socket session bounds stalled read-receipt operations", async () => {
     const onMessage = vi.fn(async () => undefined);
     const logSpy = vi.spyOn(defaultRuntime, "log").mockImplementation(() => undefined);
     const { listener, sock } = await startInboxMonitor(onMessage as InboxOnMessage, {
@@ -1713,7 +1713,7 @@ describe("web monitor inbox", () => {
     }
   });
 
-  it("bounds reconnect-gap retries even when reconnect attempts are unlimited", async () => {
+  it("socket session bounds reconnect-gap retries when attempts are unlimited", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRef = createSocketRef();
     const { listener, inbound } = await primeInboundReplyHandle({
@@ -1741,7 +1741,7 @@ describe("web monitor inbox", () => {
   // of throwing RECONNECT_IN_PROGRESS. Before the registry fallback was added,
   // sendTrackedMessage only knew its own controller's socketRef and the captured
   // reply was permanently broken.
-  it("routes the captured reply through a successor controller when the original controller's socket is gone", async () => {
+  it("socket session routes captured replies through a matching successor", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRefA = createSocketRef();
 
@@ -1807,7 +1807,7 @@ describe("web monitor inbox", () => {
   // `<x>@lid`). The fallback must recognize the same account when one side
   // reports only PN and the other only LID, because the captured reply
   // shouldn't be dropped just because the identity form rotated.
-  it("accepts successor-controller fallback when only the LID-vs-PN form differs for the same account e164", async () => {
+  it("socket session accepts a successor with equivalent LID and phone identity", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRefA = createSocketRef();
 
@@ -1866,7 +1866,7 @@ describe("web monitor inbox", () => {
   // VE-513 session-safety guard: if the registered successor controller has been
   // re-linked to a different WhatsApp identity (different self JID), the
   // captured reply must fail closed rather than route through the wrong number.
-  it("refuses successor-controller fallback when the registered controller's self JID does not match", async () => {
+  it("socket session refuses a successor with a mismatched self identity", async () => {
     const onMessage = vi.fn(async () => undefined);
     const socketRefA = createSocketRef();
